refactor: put PrimaryGuild in user.py
