+++ nextcord/__init__.py
@@ -51,7 +51,6 @@ from .object import *
 from .partial_emoji import *
 from .permissions import *
 from .player import *
-from .primary_guild import *
 from .raw_models import *
 from .reaction import *
 from .role import *
